update working middleware
